+++ src/main/java/com/example/demo/repository/ChildRepository.java
@@ -70,4 +70,9 @@ public interface ChildRepository extends JpaRepository<Child, Long> {
     @Transactional
     @Query("UPDATE Child c SET c.name = ?2, c.birthYear = ?3 WHERE c.id = ?1")
     int updateChild(long id, String name, LocalDate birthYear);
+
+    @Modifying
+    @Transactional
+    @Query("DELETE FROM Child c WHERE c.id = ?1")
+    int deleteById(long id);
 }
